+++ src/patterns/ActionPanels/index.js
@@ -5,4 +5,4 @@ export { default as Panel1 } from './panel1';
 export { default as Panel2 } from './panel2';
 export { default as Panel3 } from './panel3';
 export { default as Panel4 } from './panel4';
-export { default as Panel5 } from './panel5';
+export { default as Panel7 } from './panel7';
